+++ core/src/main/java/org/apache/mahout/cf/taste/impl/common/FullRunningAverage.java
@@ -20,33 +20,41 @@ package org.apache.mahout.cf.taste.impl.common;
 import java.io.Serializable;
 
 /**
- * <p>A simple class that can keep track of a running avearage of a series of numbers. One can add to or remove from the
- * series, as well as update a datum in the series. The class does not actually keep track of the series of values, just
- * its running average, so it doesn't even matter if you remove/change a value that wasn't added.</p>
+ * <p>
+ * A simple class that can keep track of a running avearage of a series of numbers. One can add to or remove
+ * from the series, as well as update a datum in the series. The class does not actually keep track of the
+ * series of values, just its running average, so it doesn't even matter if you remove/change a value that
+ * wasn't added.
+ * </p>
  */
 public class FullRunningAverage implements RunningAverage, Serializable {
-
+  
   private int count;
   private double average;
-
+  
   public FullRunningAverage() {
     count = 0;
     average = Double.NaN;
   }
-
-  /** @param datum new item to add to the running average */
+  
+  /**
+   * @param datum
+   *          new item to add to the running average
+   */
   @Override
   public synchronized void addDatum(double datum) {
     if (++count == 1) {
       average = datum;
     } else {
-      average = average * ((double) (count - 1) / (double) count) + datum / (double) count;
+      average = average * (count - 1) / count + datum / count;
     }
   }
-
+  
   /**
-   * @param datum item to remove to the running average
-   * @throws IllegalStateException if count is 0
+   * @param datum
+   *          item to remove to the running average
+   * @throws IllegalStateException
+   *           if count is 0
    */
   @Override
   public synchronized void removeDatum(double datum) {
@@ -56,35 +64,37 @@ public class FullRunningAverage implements RunningAverage, Serializable {
     if (--count == 0) {
       average = Double.NaN;
     } else {
-      average = average * ((double) (count + 1) / (double) count) - datum / (double) count;
+      average = average * (count + 1) / count - datum / count;
     }
   }
-
+  
   /**
-   * @param delta amount by which to change a datum in the running average
-   * @throws IllegalStateException if count is 0
+   * @param delta
+   *          amount by which to change a datum in the running average
+   * @throws IllegalStateException
+   *           if count is 0
    */
   @Override
   public synchronized void changeDatum(double delta) {
     if (count == 0) {
       throw new IllegalStateException();
     }
-    average += delta / (double) count;
+    average += delta / count;
   }
-
+  
   @Override
   public synchronized int getCount() {
     return count;
   }
-
+  
   @Override
   public synchronized double getAverage() {
     return average;
   }
-
+  
   @Override
   public synchronized String toString() {
     return String.valueOf(average);
   }
-
+  
 }
