+++ core/src/main/java/org/apache/mahout/cf/taste/impl/common/FullRunningAverageAndStdDev.java
@@ -17,35 +17,39 @@
 
 package org.apache.mahout.cf.taste.impl.common;
 
-/** <p>Extends {@link FullRunningAverage} to add a running standard deviation computation.</p> */
+/**
+ * <p>
+ * Extends {@link FullRunningAverage} to add a running standard deviation computation.
+ * </p>
+ */
 public final class FullRunningAverageAndStdDev extends FullRunningAverage implements RunningAverageAndStdDev {
-
+  
   private double stdDev;
   private double sumX2;
-
+  
   public FullRunningAverageAndStdDev() {
     stdDev = Double.NaN;
   }
-
+  
   @Override
   public synchronized double getStandardDeviation() {
     return stdDev;
   }
-
+  
   @Override
   public synchronized void addDatum(double datum) {
     super.addDatum(datum);
     sumX2 += datum * datum;
     recomputeStdDev();
   }
-
+  
   @Override
   public synchronized void removeDatum(double datum) {
     super.removeDatum(datum);
     sumX2 -= datum * datum;
     recomputeStdDev();
   }
-
+  
   /**
    * @throws UnsupportedOperationException
    */
@@ -53,20 +57,20 @@ public final class FullRunningAverageAndStdDev extends FullRunningAverage implem
   public void changeDatum(double delta) {
     throw new UnsupportedOperationException();
   }
-
+  
   private synchronized void recomputeStdDev() {
     int count = getCount();
     if (count > 1) {
       double average = getAverage();
-      stdDev = Math.sqrt((sumX2 - average * average * (double) count) / (double) (count - 1));
+      stdDev = Math.sqrt((sumX2 - average * average * count) / (count - 1));
     } else {
       stdDev = Double.NaN;
     }
   }
-
+  
   @Override
   public synchronized String toString() {
     return String.valueOf(String.valueOf(getAverage()) + ',' + stdDev);
   }
-
+  
 }
